# api-docs.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "Attendly API",
        "description": "교회 출석 관리 시스템 API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Generated server url"
        }
    ],
    "security": [
        {
            "bearerAuth": []
        }
    ],
...
    "paths": {
        "/api/admin/users/{userId}": {
            "get": {
                "tags": [
                    "관리자-사용자"
                ],
                "summary": "사용자 조회",
                "description": "특정 사용자 정보를 조회합니다",
                "operationId": "getUser",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseUserResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "관리자-사용자"
                ],
                "summary": "사용자 수정",
                "description": "사용자 정보를 수정합니다",
                "operationId": "updateUser",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserUpdateRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseUserResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "관리자-사용자"
                ],
                "summary": "사용자 삭제",
                "description": "사용자를 삭제합니다",
                "operationId": "deleteUser",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseVoid"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/auth/signup": {
            "post": {
                "tags": [
                    "인증 API"
                ],
                "summary": "회원가입",
                "description": "새로운 사용자를 등록합니다.",
                "operationId": "signup",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SignupRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseSignupResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/auth/refresh": {
            "post": {
                "tags": [
                    "인증 API"
                ],
                "summary": "토큰 갱신",
                "description": "리프레시 토큰을 사용하여 새로운 엑세스 토큰과 리프레시 토큰을 발급받습니다.",
                "operationId": "refreshToken",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TokenRefreshRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseTokenRefreshResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/auth/login": {
            "post": {
                "tags": [
                    "인증 API"
                ],
                "summary": "로그인",
                "description": "이메일과 비밀번호로 로그인하여 JWT 토큰을 발급받습니다.",
                "operationId": "login",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseLoginResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/delegations": {
            "post": {
                "tags": [
                    "leader-delegation-controller"
                ],
                "operationId": "createDelegation",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/DelegationCreateRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseLeaderDelegationResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/gbs-members/{gbsId}": {
            "get": {
                "tags": [
                    "GBS 멤버 관리"
                ],
                "summary": "GBS 멤버 조회",
                "description": "특정 GBS에 속한 멤버들의 정보를 조회합니다. 리더나 위임받은 리더만 자신의 GBS를 조회할 수 있습니다.",
                "operationId": "getGbsMembers",
                "parameters": [
                    {
                        "name": "gbsId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "date",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseGbsMembersListResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/gbs-members/leaders/{leaderId}/history": {
            "get": {
                "tags": [
                    "GBS 멤버 관리"
                ],
                "summary": "리더의 GBS 히스토리 조회",
                "description": "특정 리더가 지금까지 참여했던 GBS 히스토리 정보를 조회합니다. 각 GBS 히스토리에는 참여 기간과 조원 정보가 포함됩니다.",
                "operationId": "getLeaderGbsHistories",
                "parameters": [
                    {
                        "name": "leaderId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseLeaderGbsHistoryListResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/delegations/active": {
            "get": {
                "tags": [
                    "leader-delegation-controller"
                ],
                "operationId": "getActiveDelegations",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "date",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponsePageResponseLeaderDelegationResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "UserUpdateRequest": {
                "required": [
                    "name"
                ],
                "type": "object",
                "properties": {
                    "name": {
                        "maxLength": 50,
                        "minLength": 2,
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "phoneNumber": {
                        "pattern": "^\\d{3}-?\\d{3,4}-?\\d{4}$",
                        "type": "string"
                    },
                    "birthDate": {
                        "type": "string",
                        "format": "date"
                    },
                    "role": {
                        "type": "string",
                        "enum": [
                            "ADMIN",
                            "MINISTER",
                            "VILLAGE_LEADER",
                            "LEADER",
                            "MEMBER"
                        ]
                    },
                    "departmentId": {
                        "type": "integer",
                        "format": "int64"
                    }
                }
            },
            "ApiResponseUserResponse": {
                "required": [
                    "code",
                    "message",
                    "success",
                    "timestamp"
                ],
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean"
                    },
                    "timestamp": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "data": {
                        "$ref": "#/components/schemas/UserResponse"
                    },
                    "message": {
                        "type": "string"
                    },
                    "code": {
                        "type": "integer",
                        "format": "int32"
                    }
                }
            },
...
            "SignupRequest": {
                "required": [
                    "departmentId",
                    "email",
                    "name",
                    "password",
                    "role"
                ],
                "type": "object",
                "properties": {
                    "email": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "phoneNumber": {
                        "pattern": "^\\d{3}-?\\d{3,4}-?\\d{4}$",
                        "type": "string"
                    },
                    "role": {
                        "type": "string",
                        "enum": [
                            "ADMIN",
                            "MINISTER",
                            "VILLAGE_LEADER",
                            "LEADER",
                            "MEMBER"
                        ]
                    },
                    "departmentId": {
                        "type": "integer",
                        "format": "int64"
                    }
                }
            },
            "ApiResponseSignupResponse": {
                "required": [
                    "code",
                    "message",
                    "success",
                    "timestamp"
                ],
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean"
                    },
                    "timestamp": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "data": {
                        "$ref": "#/components/schemas/SignupResponse"
                    },
                    "message": {
                        "type": "string"
                    },
                    "code": {
                        "type": "integer",
                        "format": "int32"
                    }
                }
            },
            "SignupResponse": {
                "required": [
                    "email",
                    "name",
                    "role",
                    "userId"
                ],
                "type": "object",
                "properties": {
                    "userId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "name": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "phoneNumber": {
                        "type": "string"
                    },
                    "role": {
                        "type": "string"
                    }
                }
            },
            "TokenRefreshRequest": {
                "required": [
                    "refreshToken"
                ],
                "type": "object",
                "properties": {
                    "refreshToken": {
                        "type": "string"
                    }
                }
            },
            "ApiResponseTokenRefreshResponse": {
                "required": [
                    "code",
                    "message",
                    "success",
                    "timestamp"
                ],
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean"
                    },
                    "timestamp": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "data": {
                        "$ref": "#/components/schemas/TokenRefreshResponse"
                    },
                    "message": {
                        "type": "string"
                    },
                    "code": {
                        "type": "integer",
                        "format": "int32"
                    }
                }
            },
...
            "LoginRequest": {
                "required": [
                    "email",
                    "password"
                ],
                "type": "object",
                "properties": {
                    "email": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    }
                }
            },
            "ApiResponseLoginResponse": {
                "required": [
                    "code",
                    "message",
                    "success",
                    "timestamp"
                ],
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean"
                    },
                    "timestamp": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "data": {
                        "$ref": "#/components/schemas/LoginResponse"
                    },
                    "message": {
                        "type": "string"
                    },
                    "code": {
                        "type": "integer",
                        "format": "int32"
                    }
                }
            },
...
            "DelegationCreateRequest": {
                "required": [
                    "delegateId",
                    "delegatorId",
                    "endDate",
                    "gbsGroupId",
                    "startDate"
                ],
                "type": "object",
                "properties": {
                    "delegatorId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "delegateId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "gbsGroupId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "startDate": {
                        "type": "string",
                        "format": "date"
                    },
                    "endDate": {
                        "type": "string",
                        "format": "date"
                    }
                }
            },
            "ApiResponseLeaderDelegationResponse": {
                "required": [
                    "code",
                    "message",
                    "success",
                    "timestamp"
                ],
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean"
                    },
                    "timestamp": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "data": {
                        "$ref": "#/components/schemas/LeaderDelegationResponse"
                    },
                    "message": {
                        "type": "string"
                    },
                    "code": {
                        "type": "integer",
                        "format": "int32"
                    }
                }
            },
...
            "ApiResponseVoid": {
                "required": [
                    "code",
                    "message",
                    "success",
                    "timestamp"
                ],
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean"
                    },
                    "timestamp": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "data": {
                        "type": "object"
                    },
                    "message": {
                        "type": "string"
                    },
                    "code": {
                        "type": "integer",
                        "format": "int32"
                    }
                }
            },
...
            "ApiResponseGbsMembersListResponse": {
                "required": [
                    "code",
                    "message",
                    "success",
                    "timestamp"
                ],
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean"
                    },
                    "timestamp": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "data": {
                        "$ref": "#/components/schemas/GbsMembersListResponse"
                    },
                    "message": {
                        "type": "string"
                    },
                    "code": {
                        "type": "integer",
                        "format": "int32"
                    }
                }
            },
...
            "ApiResponseLeaderGbsHistoryListResponse": {
                "required": [
                    "code",
                    "message",
                    "success",
                    "timestamp"
                ],
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean"
                    },
                    "timestamp": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "data": {
                        "$ref": "#/components/schemas/LeaderGbsHistoryListResponse"
                    },
                    "message": {
                        "type": "string"
                    },
                    "code": {
                        "type": "integer",
                        "format": "int32"
                    }
                }
            },
...
            "ApiResponsePageResponseLeaderDelegationResponse": {
                "required": [
                    "code",
                    "message",
                    "success",
                    "timestamp"
                ],
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean"
                    },
                    "timestamp": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "data": {
                        "$ref": "#/components/schemas/PageResponseLeaderDelegationResponse"
                    },
                    "message": {
                        "type": "string"
                    },
                    "code": {
                        "type": "integer",
                        "format": "int32"
                    }
                }
            },
...
        },
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}